fixed small bug in login, new customError class to allow easy return of error codes with custom errors
