--- backend/validation/FormatError.js
@@ -1,9 +1,0 @@
-class FormatError extends Error {
-  constructor(error) {
-    super(error);
-    this.name = "FormatError";
-    this.code = 400;
-  }
-}
-
-module.exports = FormatError;
